Splitting out model-specific source handling #13
This mostly finishes the initial version of the modular sources refactor.
