+++ app/Jobs/ProcessChannelImport.php
@@ -13,6 +13,7 @@ class ProcessChannelImport implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
+    protected $type;
     protected $channelId;
     protected $importVideos;
     protected $importPlaylists;
@@ -22,8 +23,9 @@ class ProcessChannelImport implements ShouldQueue
      *
      * @return void
      */
-    public function __construct(string $channelId, bool $importVideos = false, bool $importPlaylists = false)
+    public function __construct(string $type, string $channelId, bool $importVideos = false, bool $importPlaylists = false)
     {
+        $this->type = $type;
         $this->channelId = $channelId;
         $this->importVideos = $importVideos;
         $this->importPlaylists = $importPlaylists;
@@ -36,6 +38,10 @@ class ProcessChannelImport implements ShouldQueue
      */
     public function handle()
     {
-        Channel::importYouTube($this->channelId, $this->importVideos, $this->importPlaylists);
+        if ($this->type == 'youtube') {
+            Channel::importYouTube($this->channelId, $this->importVideos, $this->importPlaylists);
+        } else {
+            Channel::import($this->type, $this->channelId);
+        }
     }
 }
