+++ resources/views/sources/embed-twitter.blade.php
@@ -1,0 +1,7 @@
+<div class="mb-4">
+    <blockquote class="twitter-tweet" data-conversation="none" data-align="center" data-dnt="true">
+        <p>{{ $video->description }}</p>
+        <a href="{{ $video->source_link }}">{{ $video->published_at }}</a>
+    </blockquote>
+    <script async defer src="https://platform.twitter.com/widgets.js"></script>
+</div>
