+++ resources/views/sources/embed-youtube.blade.php
@@ -1,0 +1,7 @@
+<div class="relative mb-4 lg:mb-6 pb-9/16">
+    <iframe class="absolute w-full h-full"
+        src="https://www.youtube.com/embed/{{ $video->uuid }}"
+        referrerpolicy="no-referrer"
+        allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+        allowfullscreen></iframe>
+</div>
